cgi-detector-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
Pillow==10.1.0
python-multipart==0.0.6
scikit-image==0.22.0
PyWavelets==1.5.0